refactor(structure): 백엔드 도메인 패키지 구조 정리
