+++ backend/src/main/java/com/back/coach/domain/roadmap/dto/RoadmapProgressCommandResult.java
@@ -1,0 +1,13 @@
+package com.back.coach.domain.roadmap.dto;
+
+import com.back.coach.global.code.ProgressStatus;
+
+import java.time.Instant;
+
+public record RoadmapProgressCommandResult(
+        Long progressLogId,
+        Long roadmapWeekId,
+        ProgressStatus status,
+        Instant savedAt
+) {
+}
