+++ backend/src/main/java/com/back/coach/domain/roadmap/dto/RoadmapProgressRequest.java
@@ -1,0 +1,12 @@
+package com.back.coach.domain.roadmap.dto;
+
+import com.back.coach.global.code.ProgressStatus;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+
+public record RoadmapProgressRequest(
+        @NotNull Long roadmapWeekId,
+        @NotNull ProgressStatus status,
+        @Size(max = 1000) String note
+) {
+}
